+++ .flake8
@@ -1,8 +1,7 @@
 [flake8]
 max-line-length = 88
-ignore = E501, W503, E266 
+ignore = E501, W503, E266, E203
 #E501 line too long
 #W503 line break before binary operator
 ; exclude = .git,__pycache__,docs
 max-complexity = 15
-extend-ignore = 203
