+++ goblet/handlers/bq_remote_function.py
@@ -4,7 +4,7 @@ from typing import get_type_hints
 
 from googleapiclient.errors import HttpError
 
-from goblet.resources.handler import Handler
+from goblet.handlers.handler import Handler
 from goblet_gcp_client.client import get_default_project
