+++ goblet/handlers/eventarc.py
@@ -6,7 +6,7 @@ from goblet.config import GConfig
 from goblet.response import Response
 import logging
 
-from goblet.resources.handler import Handler
+from goblet.handlers.handler import Handler
 from goblet_gcp_client.client import get_default_project, get_default_location
 
 log = logging.getLogger("goblet.deployer")
