+++ goblet/handlers/storage.py
@@ -1,7 +1,7 @@
 from goblet.config import GConfig
 import logging
 
-from goblet.resources.handler import Handler
+from goblet.handlers.handler import Handler
 from goblet_gcp_client.client import get_default_project, get_default_location
 from goblet.common_cloud_actions import (
     get_function_runtime,
